requirements.txt
yfinance==0.2.36
pandas==2.2.0
SQLAlchemy==2.0.27
python-dotenv==1.0.1
schedule==1.2.1
requests==2.31.0 